Clean up e2e server security tests
Moved with:
https://github.com/product-os/jellyfish-action-library/pull/1301
https://github.com/product-os/jellyfish-plugin-default/pull/703

Change-type: patch
